fix(test): support generated blank IDs

+++ test-service/src/main/java/vn/edu/primary/test/dto/BlankDTO.java
@@ -10,7 +10,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class BlankDTO {
-    private Integer id;
+    // The web editor uses Date.now() for newly added blanks, which exceeds
+    // Integer.MAX_VALUE. Keep this as Long so the request can be deserialized.
+    private Long id;
     private Integer position;
     private String correctAnswer;
     private Integer points;
